refactor: Repository 인터페이스 도메인 패키지로 이동 개선
