+++ src/test/java/roomescape/repository/JdbcReservationTimeRepositoryTest.java
@@ -17,6 +17,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.context.jdbc.Sql.ExecutionPhase;
 import roomescape.domain.ReservationTime;
+import roomescape.domain.ReservationTimeRepository;
 import roomescape.dto.response.ReservationTimeWithBookStatusResponse;
 
 // 테스트 초기 데이터
