+++ src/test/java/roomescape/service/FakeReservationRepository.java
@@ -8,9 +8,9 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicLong;
 import roomescape.domain.Reservation;
+import roomescape.domain.ReservationRepository;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
-import roomescape.repository.ReservationRepository;
 
 public class FakeReservationRepository implements ReservationRepository {
