+++ src/test/java/roomescape/service/FakeThemeRepository.java
@@ -7,7 +7,7 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicLong;
 import roomescape.domain.Theme;
-import roomescape.repository.ThemeRepository;
+import roomescape.domain.ThemeRepository;
 
 public class FakeThemeRepository implements ThemeRepository {
